# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Momento Analytics API
  version: 0.5.0
  description: On-premise event collection, reporting, management, and query API.
servers:
  - url: /
paths:
  /collect/v1/events:
    post:
      summary: Accept browser or server events into the durable inbox
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CollectRequest'}
      responses:
        '202': {description: Durable acceptance completed}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /api/v1/auth/login:
    post:
      summary: Local administrator login
      security: []
      responses: {'200': {description: Session cookie issued}, '401': {$ref: '#/components/responses/Error'}}
  /api/v1/sites:
    get:
      summary: List analytics sites
      responses: {'200': {description: Site list}}
    post:
      summary: Create a site and issue its tracking key once
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SiteMutation'}
      responses: {'201': {description: Site created}}
  /api/v1/sites/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
    patch:
      summary: Update site timezone, engagement threshold, domains, and collection settings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SiteMutation'}
      responses: {'200': {description: Site updated}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/sites/{siteId}/overview:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: Dashboard metrics and daily trend
      responses: {'200': {description: Overview}}
  /api/v1/sites/{siteId}/realtime:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: Last 1, 5, and 30 minute activity
      responses: {'200': {description: Realtime report}}
  /api/v1/sites/{siteId}/usage:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: Internal organization, department, service, feature, button, and network usage
      responses: {'200': {description: Internal usage report}}
  /api/v1/sites/{siteId}/sessions:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List materialized session summaries
      responses: {'200': {description: Session rows}}
  /api/v1/sites/{siteId}/identities:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List deterministic User ID to Visitor ID identity links
      description: Disabled when the Visitor Profile privacy policy is disabled.
      responses: {'200': {description: Identity graph rows}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/sites/{siteId}/ecommerce:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: Revenue, commerce funnel, and product performance
      responses: {'200': {description: Ecommerce report}}
  /api/v1/sites/{siteId}/visitors/{visitorId}/timeline:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {name: visitorId, in: path, required: true, schema: {type: string}}
    get:
      summary: Visitor event timeline; disabled when visitor profiles are disabled
      responses: {'200': {description: Visitor timeline}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/sites/{siteId}/retention:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: Read the site retention policy
      responses: {'200': {description: Retention policy}}
    put:
      summary: Replace the site retention policy
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RetentionPolicy'}
      responses: {'200': {description: Updated}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/segments:
    get:
      summary: List visible saved segments
      parameters: [{name: site_id, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Segment list}}
    post:
      summary: Create a nested AND/OR segment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SegmentSaveRequest'}
      responses: {'201': {description: Segment created}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/segments/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
    put:
      summary: Update a segment owned by the user or workspace
      responses: {'200': {description: Updated}}
    delete:
      summary: Delete a segment
      responses: {'204': {description: Deleted}}
  /api/v1/dimensions:
    get:
      summary: List the custom dimension registry
      parameters: [{name: site_id, in: query, required: true, schema: {type: string}}]
      responses: {'200': {description: Dimension list}}
    post:
      summary: Register or update a custom dimension
      responses: {'200': {description: Dimension saved}}
  /api/v1/reports:
    get:
      summary: List saved explorations and reports
      parameters:
        - {name: site_id, in: query, required: true, schema: {type: string}}
        - {name: kind, in: query, schema: {type: string}}
      responses: {'200': {description: Saved report list}}
    post:
      summary: Save an exploration, funnel, path, or dashboard definition
      responses: {'201': {description: Report saved}}
  /api/v1/query:
    post:
      summary: Run an allowlisted dimension and metric query
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/QueryRequest'}
      responses: {'200': {description: Query rows}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/privacy/delete:
    post:
      summary: Delete analytics data and atomically reconcile queue payloads, Raw Events, identity links, Visitor/Session summaries, and daily aggregates
      responses: {'200': {description: Data deleted or property scrubbed}, '400': {$ref: '#/components/responses/Error'}}
...
  /api/v1/sites/{siteId}/environments:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List DEV/STG/PRD and custom collection environments
      responses: {'200': {description: Environment policies}}
  /api/v1/sites/{siteId}/environments/{name}:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {name: name, in: path, required: true, schema: {type: string}}
    put:
      summary: Create or update contract mode and cardinality limit for an environment
      responses: {'200': {description: Environment updated}}
  /api/v1/sites/{siteId}/event-contracts:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List versioned event contracts
      responses: {'200': {description: Contract versions}}
    post:
      summary: Create a draft or active event contract version
      responses: {'201': {description: Contract version created}}
  /api/v1/sites/{siteId}/event-contracts/{eventName}/{version}/activate:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {name: eventName, in: path, required: true, schema: {type: string}}
      - {name: version, in: path, required: true, schema: {type: integer}}
    post:
      summary: Activate an immutable event contract version
      responses: {'200': {description: Contract activated}}
...
  /api/v1/sites/{siteId}/data-quality:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Tracking health, contract, PII, lag, duplicate and cardinality report
      responses: {'200': {description: Data Quality Center report}}
  /api/v1/sites/{siteId}/cohort:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Cohort and retention matrix
      responses: {'200': {description: Retention cohorts}}
  /api/v1/sites/{siteId}/journeys:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List personal and shared Business Journeys
      responses: {'200': {description: Journey definitions}}
    post:
      summary: Save a 2-12 step Business Journey
      responses: {'200': {description: Journey saved}}
  /api/v1/sites/{siteId}/journeys/analyze:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    post:
      summary: Analyze sequential business outcomes with a conversion window
      responses: {'200': {description: Journey conversion and elapsed time}}
  /api/v1/sites/{siteId}/adoption:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Organization, department and feature adoption
      responses: {'200': {description: Adoption, repeat use and dormant users}}
  /api/v1/sites/{siteId}/experience:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Core Web Vitals, errors, conversion impact and release impact
      responses: {'200': {description: Experience analytics}}
  /api/v1/sites/{siteId}/ai-analytics:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Model, Agent, MCP and Tool usage, latency, tokens and cost
      responses: {'200': {description: AI operations analytics}}
  /api/v1/sites/{siteId}/insights:
    parameters:
      - {$ref: '#/components/parameters/SiteId'}
      - {$ref: '#/components/parameters/Environment'}
    get:
      summary: Offline rule-based change detection and recommended diagnosis
      responses: {'200': {description: Insights}}
  /api/v1/sites/{siteId}/natural-query:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    post:
      summary: Ask Korean or English analytics questions with the offline semantic parser
      responses: {'200': {description: Interpreted intent, answer and evidence}}
...
  /api/v1/sites/{siteId}/scheduled-reports:
    parameters: [{$ref: '#/components/parameters/SiteId'}]
    get:
      summary: List scheduled reports and Segment actions
      responses: {'200': {description: Schedules}}
    post:
      summary: Schedule report or Segment aggregate delivery
      responses: {'200': {description: Schedule saved}}
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Personal mom_key_ API key or browser session token
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      schema: {type: string}
    Environment:
      name: environment
      in: query
      required: false
      schema: {type: string, default: prd}
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties: {code: {type: string}, message: {type: string}}
  schemas:
    CollectRequest:
      type: object
      required: [site_id, visitor_id, session_id, events]
      properties:
        site_id: {type: string}
        environment: {type: string, default: prd, pattern: '^[a-z][a-z0-9_-]{0,31}$'}
        tracking_key: {type: string}
        visitor_id: {type: string, maxLength: 128}
        session_id: {type: string, maxLength: 128}
        user_id: {type: string, maxLength: 128}
        user_properties: {type: object, additionalProperties: true}
        context: {$ref: '#/components/schemas/EventContext'}
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [name, timestamp]
            properties:
              id: {type: string, format: uuid}
              name: {type: string, pattern: '^[A-Za-z][A-Za-z0-9_]{0,63}$'}
              timestamp: {type: integer}
              contract_version: {type: integer, minimum: 1, default: 1}
              properties: {type: object, additionalProperties: true}
              context:
                allOf:
                  - {$ref: '#/components/schemas/EventContext'}
                description: Event occurrence-time snapshot; overrides the request-level fallback context.
...
security:
  - bearer: []